Move flavor traits
